+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 ## 1.1.0 (In development)
+### New Features
+* Items can now be removed via index with `ListView.removeAt`
+
+### Bug Fixes
 * Restored `Listview.remove` after its removal in the 1.0.0 refactor
 
 ## 1.0.0 (January 22, 2019)